# abei/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "fireflyToken": {
        "description": "Firefly III 的个人访问令牌，原样透传。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "资源接口与能力目录。账本操作在过渡期委托 Firefly III。",
    "title": "阿贝 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/bills/process": {
      "post": {
        "description": "立刻推进一轮账单解析，把待处理的任务往前走一步。",
        "operationId": "bills.process",
        "parameters": [
          {
            "description": "只跑校验和预览，不落库。响应会带 dry_run: true。",
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "显式确认。这条能力是 draft 档，不带也能执行。",
            "in": "query",
            "name": "confirm",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "additionalProperties": false,
                "description": "`bills sync` 与 `bills process` 的参数。",
                "properties": {
                  "limit": {
                    "description": "这一轮最多处理多少封，默认 100。",
                    "format": "uint32",
                    "minimum": 0,
                    "type": [
                      "integer",
                      "null"
                    ]
                  }
                },
                "title": "BillsBatchParams",
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "立刻推进一轮账单解析，把待处理的任务往前走一步。"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          }
        },
        "summary": "解析账单",
        "tags": [
          "bills"
        ],
        "x-abei-backend": "firefly",
        "x-abei-risk": "draft"
      }
    },
    "/v1/bills/sync": {
      "post": {
        "description": "提交一次邮箱同步任务；任务在后台拉取并解析新账单邮件。",
        "operationId": "bills.sync",
        "parameters": [
          {
            "description": "只跑校验和预览，不落库。响应会带 dry_run: true。",
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "显式确认。这条能力是 draft 档，不带也能执行。",
            "in": "query",
            "name": "confirm",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "additionalProperties": false,
                "description": "`bills sync` 与 `bills process` 的参数。",
                "properties": {
                  "limit": {
                    "description": "这一轮最多处理多少封，默认 100。",
                    "format": "uint32",
                    "minimum": 0,
                    "type": [
                      "integer",
                      "null"
                    ]
                  }
                },
                "title": "BillsBatchParams",
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "提交一次邮箱同步任务；任务在后台拉取并解析新账单邮件。"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "提交一次邮箱同步任务；任务在后台拉取并解析新账单邮件。"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "reason": {
                      "description": "机读错误码，驼峰。",
                      "type": "string"
                    },
                    "resource": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "upstream": {},
                    "verb": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "reason"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "RFC 9457 problem+json。reason 是机读驼峰码。"
          }
        },
        "summary": "收取账单邮件",
        "tags": [
          "bills"
        ],
        "x-abei-backend": "server",
        "x-abei-risk": "draft"
      }
    },
...
  },
  "security": [
    {
      "fireflyToken": []
    }
  ],
  "servers": [
    {
      "description": "当前实例",
      "url": "/"
    }
  ]
}